FIX: InlineOneboxer watched word censor error (#16921)
In 7328a2bfb056915efc1e42dd1eea64da28b11f12 we changed the
InlineOneboxer#onebox_for method to run the title of the
onebox through WatchedWord#censor_text. However, it is
allowable for the title to be nil, which was causing this
error in production:

> NoMethodError : undefined method gsub for nil:NilClass

We just need to check whether the title is nil before trying
to censor it.